Bring support for MIPS 32 big endian
Feature:
    - Add CONFIG_32BIT and CONFIG_64BIT for mips32 and mips64el

To fix:
    - Revert commit b5b1c4ec4506df7cee6a9ba8ffff36f43e0cd8e3
      musl libc 1.1.14 does not has support for syscalls fsopen, fsconfig and fsmount
    - Disable gnutls and selinux recquirments

Note:
    - When compel converts the compiled parasite and restorer to a C array, it does so by parsing the object.
      An assumption on similar endianness between the host and the target is made.                                                                                                           14     A workaround consists to do the relocation on the target before continuing compiling the rest of CRIU.

+++ criu/arch/mips/restorer.c
@@ -23,9 +23,9 @@ unsigned long arch_shmat(int shmid, void *shmaddr, int shmflg, unsigned long siz
 
 	/* SHMLBA-aligned, direct call shmat() */
 	if (!((unsigned long)shmaddr & (SHMLBA - 1)))
-		return sys_shmat(shmid, shmaddr, shmflg);
+		return (unsigned long) sys_shmat(shmid, shmaddr, shmflg);
 
-	smap = sys_shmat(shmid, NULL, shmflg);
+	smap = (unsigned long) sys_shmat(shmid, NULL, shmflg);
 	if (IS_ERR_VALUE(smap)) {
 		pr_err("shmat() with NULL shmaddr failed: %d\n", (int)smap);
 		return smap;
@@ -40,7 +40,7 @@ unsigned long arch_shmat(int shmid, void *shmaddr, int shmflg, unsigned long siz
 	pr_warn("Make sure that you don't migrate shmem from non-VIPT cached CPU to VIPT cached \n");
 	pr_warn("Otherwise YOU HAVE A CHANCE OF DATA CORRUPTIONS in writeable shmem\n");
 
-	smap = sys_mremap(smap, size, size, MREMAP_FIXED | MREMAP_MAYMOVE, (unsigned long)shmaddr);
+	smap = (unsigned long) sys_mremap(smap, size, size, MREMAP_FIXED | MREMAP_MAYMOVE, (unsigned long)shmaddr);
 	if (IS_ERR_VALUE(smap))
 		pr_err("mremap() for shmem failed: %d\n", (int)smap);
 	return smap;
